SystemUI: Prevent NPE in Dependency.get() in BrightnessController
When com.android.systemui process died for some reason, the process will
be restarted and will schedule to restart services, such as
SystemUIService and KeyguardService, if start activity BrightnessDialog
before these services, then NPE will happen, the reason is that this
activity will use Dependency which is initalized in these services.

Init Dependency in BrightnessController before using it.

Change-Id: I9eded3005af15612abdae10c38b42c4af937e2c9
CRs-Fixed: 2083957