am: Save and restore TaskRecord.isAvailable
TaskRecord.isAvailable is not persistent in the file. After reboot, it
will change from false to true and this will cause the card to appear
in the RecentsActivity because the value of "TaskRecord.isAvailable" is
true after reboot. However the card is not displayed since it's original
value is false before reboot.

The card sometimes is empty because no screenshot
have been set to the task, which make users uncomfortable.

https://issuetracker.google.com/issues/111801340

Change-Id: Ic4a42d0f3d1cea2ae11f0bfba2b7544132a60278
Signed-off-by: RickAi <yongbiaoai@gmail.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>