SystemUI: Fix NPE when DozeService is destroied
When DozeService's host is not null, it can be created successful
and DozeService is added into PluginManagerImpl instance as a
PluginListener. When DozeService is stoped, it will be removed
from PluginManagerImpl instance.

When DozeService's host is null, PluginManagerImpl instance is
also null, it will cause NPE if trying to remove PluginListener
from PluginManagerImpl instance.
In this case, add null check to avoid NPE

CRs-Fixed: 2348040
Change-Id: I96c767d9e36c84926d5fb7f9cf2dc1a1c0ce2e08