SystemUI: Modify ime navigation bar layout
This change is to fix issue: "GUI:The Accessibility button is not in
the right position". Accessibility button is placed between "home"
button and "recents" button, click the icon, no response. Set emulator's
resolution to 1920*1080 and density to 213, this problem can be reproduced.

Change menu_container's attribute from "android:layout_width="match_parent"
to "android:layout_width="@dimen/navigation_side_padding" which is in
accordance with or8.1, with this patch, accessibility button can be placed
in a reasonable position, click it, work normally.

Change-Id: I15174c6b05bc4b1f583095f60359a7fb9c514bd2
Signed-off-by: Dandan Sun <dandan.sun@nxp.com>
Reviewed-on: http://androidsource.ap.freescale.net/project/6626
Reviewed-by: zhang sanshan <sanshan.zhang@nxp.com>