Improve performance of unclipped save layers
Instead of allocating a separate renderTarget and switching
between them on each draw the new implementation follows the same
pattern that the old HWUI renderer used. The area of the layer is
now copied to a buffer on the GPU, the area is then cleared, rendered
as normal, and finally the texture is redrawn using dst_over blending.

This results in no render target switches and is considerably faster
on some hardware.

Test: CtsGraphicsTestCases, CtsUiRenderingTestCases
Bug: 119222339
Change-Id: I672d860916dea851a3bc908b5c4fad08b79309fa
Merged-In: I716aac1fc31e4c7a171373d37dee82034c01cf18
Signed-off-by: mydongistiny <jaysonedson@gmail.com>