Weather: Move statusbar weather to QS statusbar header
With the notch and limitation of icons in the status bar,there's almost
no room left for weather and is really a tight fit among others.
So lets move it to the left side of the clock in QS statusbar header.

Also implemented a method to be invoked on orientation change and
use colors that adapt based on the scrim/wallpaper behind quick settings.

Signed-off-by: ganeshi4u <ganeshvarmai4u@gmail.com>
Change-Id: Icc429f30bbf29d3232864908711105c37865e511