Move statusbar weather to QS statusbar header
With the notch and limitation of icons in the status bar,there's almost
no room left for weather and is really a tight fit among others.
So lets move it to the left side of the clock in QS statusbar header.

Also nuked the DarkIconDispatcher as we won't be needing it on qs statusbar header,
so set the WeatherText tint to white.

Signed-off-by: ganeshi4u <ganeshvarmai4u@gmail.com>
Change-Id: Icc429f30bbf29d3232864908711105c37865e511