BrightnessController: tune the slider animation duration
The brightness slider is animated at a fixed max duration of
3000ms. If the brightness ramp rate is reduced to a value that is
much lower than the default rate, the brightness slider animation
can be considerably out of sync with the screen brightness
transitions.

Override the slider animator duration scale by the ratio of the
default ramp rate to the actual rate in use. The maximum scale is
clamped to 3x. Note that the purpose is not to make the slider
perfectly in sync with the screen, as the original design was
not like that either.

Signed-off-by: Chenyang Zhong <zhongcy95@gmail.com>
Change-Id: I980a8976d1d8e6fe9c295be9a06ee305a66c9448