PixelPropsUtils: Add Prime Videos
For some reasons on some global variants of Xiaomi devices,
Prime videos is not showing full video download/streaming
capabilities, which is not the case with IN region models.
This should also address the devices where the above case
is vice-versa like raphael,raphaelin.
Add Prime Videos to the list to bypass such st*pid scenarios.

test: Prime Video -> Settings -> Stream & Download -> {D*,S*}quality
- vayu,surya: Now show full streaming/download capabilities
- bhima,karna: uneffected, still show full capabilities.
- Quality before: 0.54 GB/hr for BEST; Quality after: 2.35 GB/hr for BEST.

Change-Id: Ie4399a7cdbf2a193101893ce56a0b1777e402ba6